moved klant generic classes to the new klant-generiek module
