+++ klant-generiek/gradle.properties
@@ -1,0 +1,1 @@
+isLib = true
