+++ klant-generiek/src/main/kotlin/nl/nlportal/klant/generiek/client/OpenKlantClientConfig.kt
@@ -13,26 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ritense.portal.klant.validation
+package nl.nlportal.klant.generiek.client
 
-import javax.validation.Validation
-import javax.validation.ValidationException
-import javax.validation.Validator
+import org.springframework.boot.context.properties.ConfigurationProperties
 
-class GraphQlValidator {
-
-    private var validator: Validator? = null
-
-    fun <T> validate(payload: T?) {
-        for (violation in getValidator().validate(payload)) {
-            throw ValidationException(violation.message)
-        }
-    }
-
-    private fun getValidator(): Validator {
-        if (validator == null) {
-            validator = Validation.buildDefaultValidatorFactory().validator!!
-        }
-        return validator!!
-    }
-}
+@ConfigurationProperties(prefix = "valtimo.openklant")
+class OpenKlantClientConfig(
+    var url: String = "",
+    var clientId: String = "",
+    var secret: String = "",
+    var rsin: String = ""
+)
