+++ klant-generiek/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+nl.nlportal.klant.generiek.autoconfiguration.KlantGeneriekAutoConfiguration
