+++ klant-generiek/src/test/kotlin/nl/nlportal/klant/generiek/TestApplication.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2015-2023 Ritense BV, the Netherlands.
+ *
+ * Licensed under EUPL, Version 1.2 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nl.nlportal.klant.generiek
+
+import com.ritense.portal.core.security.OauthSecurityAutoConfiguration
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+import org.springframework.context.annotation.Bean
+import org.springframework.security.config.web.server.ServerHttpSecurity
+import org.springframework.security.web.server.SecurityWebFilterChain
+
+@SpringBootApplication(
+    exclude = [
+        OauthSecurityAutoConfiguration::class
+    ]
+)
+class TestApplication {
+    fun main(args: Array<String>) {
+        runApplication<TestApplication>(*args)
+    }
+
+    @Bean
+    fun springSecurityWebFilterChain(http: ServerHttpSecurity): SecurityWebFilterChain {
+        http
+            .csrf()
+            .disable()
+            .authorizeExchange()
+            .anyExchange()
+            .permitAll()
+        return http.build()
+    }
+}
