+++ klant/src/test/kotlin/com/ritense/portal/klant/service/impl/BurgerServiceTest.kt
@@ -18,7 +18,7 @@ package com.ritense.portal.klant.service.impl
 import com.ritense.portal.commonground.authentication.CommonGroundAuthentication
 import com.ritense.portal.commonground.authentication.JwtBuilder
 import com.ritense.portal.klant.client.OpenKlantClient
-import com.ritense.portal.klant.client.OpenKlantClientConfig
+import nl.nlportal.klant.generiek.client.OpenKlantClientConfig
 import com.ritense.portal.klant.domain.klanten.Klant
 import com.ritense.portal.klant.domain.klanten.KlantCreationRequest
 import com.ritense.portal.klant.domain.klanten.KlantUpdate
